Remove `proc-macro-crate` dependency
The dependency so the proc_macro would generate the right code even if
`pretty-error-debug` was renamed or shadowed. But it comes with a metric
ton of transitive dependencies with a lenient MSRV policy.

To keep our MSRV, we used an old `proc-macro-crate`, which is not
exactly great, either. Now we not only have a ton of dependencies, but
stale, old dependencies on top of that.

This PR removes the dependency. The derive macro will not derive wrong
code if `pretty-error-debug` was shadowed or renamed. IMHO that's fine.